Eval/pyproject.toml
[project]
name = "decomposed-reasoning-eval"
version = "0.1.0"
description = "Standalone benchmark evaluation runner using lm-eval Python API."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "lm-eval[hf,vllm,math]>=0.4.7",
  "pyyaml>=6.0.0",
  "wandb>=0.17.0",
]

[project.optional-dependencies]
dev = [
  "pyright>=1.1.390",
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.pyright]
include = ["eval_runner", "tests"]
pythonVersion = "3.10"
typeCheckingMode = "basic"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["eval_runner*"]
